+++ models/request/movement.go
@@ -1,6 +1,6 @@
 package request
 
 type MovementRequest struct {
-	TargetX int `json:"target_x"`
-	TargetY int `json:"target_y"`
+	TargetX uint16 `json:"target_x"`
+	TargetY uint16 `json:"target_y"`
 }
